OTReservation/Insert.OTRequest/Insert,OTRequest/OtRequestSurgeryDetailList,UpdateOTReservationHeader

+++ Web Apps/HIMS.API/Models/OTManagement/OTBookingRequestModel.cs
@@ -94,7 +94,7 @@ namespace HIMS.API.Models.OTManagement
         public string? SurgeryPart { get; set; }
         public string? SurgeryFromTime { get; set; }
         public string? SurgeryEndTime { get; set; }
-        public int? SurgeryDuration { get; set; }
+        public double? SurgeryDuration { get; set; }
         public string? IsPrimary { get; set; }
         public long? SurgeonId { get; set; }
         public long? AnesthetistId { get; set; }
